version 1

+++ distribution/src/main/java/com/core/service/IWxPayInfoService.java
@@ -8,6 +8,6 @@ import java.util.List;
  * Created by core on 15/11/13.
  */
 public interface IWxPayInfoService {
-    List<WxPayInfo> getPayInfoByOrderId(Integer orderId);
+    List<WxPayInfo> getPayInfoByOrderId(Long orderId);
     void insertWxPayInfo(WxPayInfo payInfo);
 }
